chore: Update dependencies and models
This commit updates the dependencies in the project, specifically the `nanoid` package in the `user.model.ts` file. It also adds a new model called `Session` in the `session.model.ts` file. These changes are made to improve the overall functionality and maintainability of the codebase.

Note: This commit message follows the conventional commit format, starting with a type (chore) and providing a concise summary of the changes made.

Login nearly finished having access token and refresh token keys issues but should fix soon :)