chore: Update dependencies and models
This commit updates the dependencies in the project, specifically the `nanoid` package in the `user.model.ts` file. It also adds a new model called `Session` in the `session.model.ts` file. These changes are made to improve the overall functionality and maintainability of the codebase.

Login complete ( Jsonwebtoken dependency had to be reverted to an older version because there was a conflict with the acess keys will investigate later )