chore: Update dependencies and add CORS support
This commit updates the dependencies in the project, specifically adding the "cors" package to enable Cross-Origin Resource Sharing (CORS) support. CORS allows the frontend application to make requests to the backend API from a different domain. This change is made to improve the overall functionality and security of the codebase.

Fixed cors issue by installing dependency.
User logged in name now displaying correctly