chore: Update dependencies and add axios and https-proxy-agent packages
This commit updates the dependencies in the project, specifically adding the "axios" and "https-proxy-agent" packages. The "axios" package is used for making HTTP requests, and the "https-proxy-agent" package is used for handling HTTPS requests through a proxy. These changes are made to improve the functionality and performance of the codebase.

Note: System status working
Had to use a proxy and ignore SSL Certificate otherwise wouldnt work ( Development only )