chore: Update dependencies and add .env to .gitignore
This commit updates the dependencies in the project, specifically adding the "axios" and "https-proxy-agent" packages. It also adds the ".env" file to the ".gitignore" to prevent it from being tracked in version control. These changes are made to improve the functionality, performance, and security of the codebase.

Note: System status working
Had to use a proxy and ignore SSL Certificate otherwise wouldn't work (Development only)