feat: Add getAllUsers function to user service
This commit adds the getAllUsers function to the user service in order to fetch all users from the database. This function will be used in the getAllUsersHandler route handler to retrieve a list of all users for the admin route. The addition of this function enhances the functionality of the application by providing an endpoint to retrieve all users.

Note: RBAC Implemented